Unifico eyeglasses.py con utils.py
